+++ perun/view/tableof/__init__.py
@@ -1,0 +1,8 @@
+"""`Table` interprets the data as a two dimensional array. The cells in the table can be limited to
+certain columns only and allow output to the file.
+"""
+
+SUPPORTED_PROFILES = ["time|memory|mixed"]
+
+
+__author__ = 'Tomas Fiedor'