+++ CHANGELOG.rst
@@ -6,6 +6,8 @@ HEAD
 
 **To be included in next release**
 
+  - fix minor bug in bounds collector (unknown collector type)
+
 1.17.3 (2020-01-09)
 -------------------